+ Fixed some points based on comments
