--- src/main/java/bank/recommendationservice/fintech/ProductType.java
@@ -1,11 +1,0 @@
-package bank.recommendationservice.fintech;
-
-public enum ProductType {
-    DEBIT ("DEBIT"),
-    CREDIT ("CREDIT"),
-    INVEST ("INVEST"),
-    SAVING ("SAVING");
-
-    ProductType(String type) {
-    }
-}
