+++ src/main/java/bank/recommendationservice/fintech/controller/GlobalExceptionHandler.java
@@ -5,25 +5,31 @@ import bank.recommendationservice.fintech.exception.RecommendationNotFoundExcept
 import bank.recommendationservice.fintech.exception.RulesNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-@ControllerAdvice
+@RestControllerAdvice
 public class GlobalExceptionHandler {
 
-    @ExceptionHandler(RulesNotFoundException.class)
-    public ResponseEntity<String> handleRulesNotFoundException(RulesNotFoundException ex) {
+    //    @ExceptionHandler(RulesNotFoundException.class)
+//    public ResponseEntity<String> handleRulesNotFoundException(RulesNotFoundException ex) {
+//        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
+//    }
+//
+//    @ExceptionHandler(RecommendationNotFoundException.class)
+//    public ResponseEntity<String> handleRecommendationNotFoundException(RecommendationNotFoundException ex) {
+//        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
+//    }
+//
+//    @ExceptionHandler(NoTransactionsFoundException.class)
+//    public ResponseEntity<String> handleNoTransactionsFoundException(NoTransactionsFoundException ex) {
+//        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
+//    }
+    @ExceptionHandler({RulesNotFoundException.class, RecommendationNotFoundException.class, NoTransactionsFoundException.class})
+    public ResponseEntity<String> handleNotFoundExceptions(RuntimeException ex) {
         return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
     }
 
-    @ExceptionHandler(RecommendationNotFoundException.class)
-    public ResponseEntity<String> handleRecommendationNotFoundException(RecommendationNotFoundException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(NoTransactionsFoundException.class)
-    public ResponseEntity<String> handleNoTransactionsFoundException(NoTransactionsFoundException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
-
-    }
 }
+
+
