+++ src/main/java/bank/recommendationservice/fintech/other/ProductType.java
@@ -1,0 +1,19 @@
+package bank.recommendationservice.fintech.other;
+
+public enum ProductType {
+    DEBIT("DEBIT"),
+    CREDIT("CREDIT"),
+    INVEST("INVEST"),
+    SAVING("SAVING");
+
+    private final String type;
+
+    ProductType(String type) {
+        this.type = type;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+}
