+++ src/test/java/bank/recommendationservice/fintech/ruleimpl/impl/DebitDepositsTotalGreaterThanWithdrawsTest.java
@@ -3,7 +3,7 @@ package bank.recommendationservice.fintech.ruleimpl.impl;
 import bank.recommendationservice.fintech.exception.NoTransactionsFoundException;
 import bank.recommendationservice.fintech.repository.RecommendationsRepository;
 import bank.recommendationservice.fintech.ruleimpl.DebitDepositsTotalGreaterThanWithdraws;
-import bank.recommendationservice.fintech.ProductType;
+import bank.recommendationservice.fintech.other.ProductType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
